Moving determination of meta information to mapping source
