+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvTimeSeriesMappingSourceIT.groovy
@@ -5,12 +5,14 @@
  */
 package edu.ie3.datamodel.io.source.csv
 
+import edu.ie3.datamodel.io.connectors.CsvFileConnector
 import edu.ie3.datamodel.io.csv.FileNamingStrategy
+import edu.ie3.datamodel.io.csv.timeseries.ColumnScheme
 import edu.ie3.datamodel.io.source.TimeSeriesMappingSource
 import spock.lang.Shared
 import spock.lang.Specification
 
-class CsvTimeSeriesMappingSourceTest extends Specification implements CsvTestDataMeta {
+class CsvTimeSeriesMappingSourceIT extends Specification implements CsvTestDataMeta {
 	@Shared
 	TimeSeriesMappingSource source
 
@@ -59,5 +61,32 @@ class CsvTimeSeriesMappingSourceTest extends Specification implements CsvTestDat
 		then:
 		actual.present
 		actual.get() == expectedUuid
+	}
+
+	def "A csv time series mapping source returns empty optional on meta information for non existing time series"() {
+		given:
+		def timeSeriesUuid = UUID.fromString("f5eb3be5-98db-40de-85b0-243507636cd5")
+
+		when:
+		def actual = source.getTimeSeriesMetaInformation(timeSeriesUuid)
+
+		then:
+		!actual.present
+	}
+
+	def "A csv time series mapping source returns correct meta information for an existing time series"() {
+		given:
+		def timeSeriesUuid = UUID.fromString("3fbfaa97-cff4-46d4-95ba-a95665e87c26")
+		def expected = new CsvFileConnector.CsvIndividualTimeSeriesMetaInformation(
+				timeSeriesUuid,
+				ColumnScheme.APPARENT_POWER,
+				"its_pq_3fbfaa97-cff4-46d4-95ba-a95665e87c26")
+
+		when:
+		def actual = source.getTimeSeriesMetaInformation(timeSeriesUuid)
+
+		then:
+		actual.present
+		actual.get() == expected
 	}
 }
